Fixes bug that could cause crashes
When using an alias with $i.. parameter without providing enough params
it would expand to subList(i, params.size). If i was larger than the
number of parameters, this would violate an invariant in subList and
cause a crash.

This bug was fixed, additionally the code now matches the functionality
of the desktop version, further bugs introduced while porting the code
have been fixed.